Monitor: Operator, Resource
This commit enhances the Monitoring System, with two
components: OperatorMonitor, ResourceMonitor.

OperatorMonitor
--------------------
The OperatorMonitor retrieves the statistics related to each
operator of the topology. It needs to collect the sum of
emitted events and the amount of processed event by each
operator instance. Afterwards, aggregates the statistics per
operator. To avoid collecting data from multiple or not
consistent window, this components adopts a PULL-approach.

ResourceMonitor
-------------------
ResourceMonitor monitors the computing resources, which
are used to execute the processing nodes.

In this implementation, ResourceMonitor retrieves the CPU
utilization of each DockerContainer and stores this
information in the Runtime repository.

New properties:
- visp.monitor.period
	(default=10000, time interval between two
	consecutive invocation of OperatorMonitor;
	expressed in ms)